Build domain.lisp for new model
The new model is more extensive than the previous one.  With these
changes we support a Theme, which consists of a set of Topics.  The
theme is then subdevided into RegionThemes, each of which represent a
region in the city (or the city as a whole).  The values, one for each
supported year, are connected through instances of Value.